+++ src/scripts/typings/graphql.d.ts
@@ -1,0 +1,8 @@
+declare module '*.graphql' {
+  /* import { DocumentNode } from 'graphql';
+
+  const value: DocumentNode; */
+  const value: String;
+
+  export default value;
+}