+++ src/scripts/views/5-UserProfile/_/MorePanel/ManageExtensionUser.js
@@ -1,0 +1,36 @@
+import { ContentBoxContent, Text } from "../../../../components/style-guide";
+
+export default class ManageExtensionUser {
+  /**
+   * @param {import(".").default} main
+   */
+  constructor(main) {
+    this.main = main;
+
+    this.Render();
+    this.BindHandlers();
+  }
+  Render() {
+    this.container = ContentBoxContent({
+      spacedBottom: true,
+      children: this.link = Text({
+        type: "a",
+        size: "small",
+        weight: "bold",
+        color: "blue-dark",
+        html: System.data.locale.userProfile.morePanel
+          .manageExtensionUser.title,
+      }),
+    });
+
+    this.main.container.append(this.container);
+  }
+  BindHandlers() {
+    this.link.addEventListener("click", this.EditUser.bind(this));
+  }
+  EditUser() {
+    System.OpenExtensionOptions({
+      editUser: this.main.main.profileData.id
+    });
+  }
+}